Organized scripts in the Enemy module
Created a new hierarchy with subdirectories within the enemy module for better organization and ease of searching
